SetBuildParameterCommandTest
